.archive/mdso-dev/common-ci-cd-automations/ansible-test/image/requirements.txt
ansible==2.8.3
python-pptx==0.6.18
